--- binance_trade_bot/__main__.py
@@ -1,4 +1,0 @@
-from .crypto_trading import main
-
-if __name__ == '__main__':
-    main()
