+++ logger.py
@@ -1,9 +1,9 @@
 import logging
 import logging.handlers
+from logging import Handler, Formatter
+from notifications import NotificationHandler
 
-from .notifications import NotificationHandler
-
-LOG_PATH = "logs/crypto_trading.log"
+LOG_PATH = "crypto_trading.log"
 
 
 class Logger:
