+++ models/scout_history.py
